feat: update values to SI
